// src/generated/api.ts
export const Element = {
    Metal: 'METAL',
    Wood: 'WOOD',
    Water: 'WATER',
    Fire: 'FIRE',
    Earth: 'EARTH'
} as const;

export type Element = typeof Element[keyof typeof Element];
// ...
export const Slot = {
    Head: 'HEAD',
    Necklace: 'NECKLACE',
    Weapon: 'WEAPON',
    Armor: 'ARMOR',
    Gloves: 'GLOVES',
    Ring: 'RING',
    Amulet: 'AMULET',
    Belt: 'BELT',
    Boots: 'BOOTS'
} as const;

export type Slot = typeof Slot[keyof typeof Slot];